requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-cors-headers==2.4.1
django-decouple==2.1
django-simple-email-confirmation==0.23
djangorestframework==3.9.1
idna==2.8
phe==1.4.0
psycopg2-binary==2.7.7
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
